+++ gen/field.go
@@ -83,6 +83,10 @@ func (f fieldGroupGenerator) Generate(g Generator) error {
 		return err
 	}
 
+	if err := f.DefineDefaultConstructor(g); err != nil {
+		return err
+	}
+
 	if err := f.ToWire(g); err != nil {
 		return err
 	}
@@ -244,6 +248,35 @@ func (f *fieldGroupGenerator) declFieldName(fs *compile.FieldSpec) (string, erro
 	return name, nil
 }
 
+func (f fieldGroupGenerator) DefineDefaultConstructor(g Generator) error {
+	var hasDefaults bool
+	for _, f := range f.Fields {
+		if f.Default != nil {
+			hasDefaults = true
+			break
+		}
+	}
+	if !hasDefaults {
+		return nil
+	}
+	return g.DeclareFromTemplate(
+		`
+		// Default_<.Name> constructs a new <.Name> struct,
+		// pre-populating any fields with defined default values. 
+		func Default_<.Name>() *<.Name> {
+			<- $v := newVar "v" ->
+			var v <.Name>
+			<- range .Fields ->
+				<- $fname := goName . ->
+				<- if .Default>
+					<$v>.<$fname> = <constantValuePtr .Default .Type>
+				<- end ->
+			<end>
+			return &v
+		}
+		`, f, TemplateFunc("constantValuePtr", ConstantValuePtr))
+}
+
 func (f fieldGroupGenerator) ToWire(g Generator) error {
 	return g.DeclareFromTemplate(
 		`
@@ -297,14 +330,17 @@ func (f fieldGroupGenerator) ToWire(g Generator) error {
 						<$i>++
 				<- else ->
 					<- if .Default ->
-						if <$f> == nil {
-							<$f> = <constantValuePtr .Default .Type>
+						<- $fval := printf "%s%s" $v $fname ->
+						<$fval> := <$f>
+						if <$fval> == nil {
+							<$fval> = <constantValuePtr .Default .Type>
 						}
 						{
+							<$wVal>, err = <toWirePtr .Type $fval>
 					<- else ->
 						if <$f> != nil {
-					<- end>
 							<$wVal>, err = <toWirePtr .Type $f>
+					<- end>
 							if err != nil {
 								return <$wVal>, err
 							}